(imp): creating Lens Compoenent to manipulate the MainCamera configs better
- just updating the FOV yet

+++ Modules/VirtualCamera/Runtime/VirtualCamera.gd
@@ -1,4 +1,7 @@
+@tool
 class_name VirtualCamera extends Node3D
+
+var _isEditorMode = Engine.is_editor_hint()
 
 @export var enabled: bool = true:
 	get:
@@ -6,8 +9,8 @@ class_name VirtualCamera extends Node3D
 	set(value):
 		if enabled == value: return
 		enabled = value
+		if _isEditorMode: return
 		VirtualCameraService.onEnabledModified.emit(self)
-		print(name, " ", "enabled ",enabled)
 
 @export var priority: int = 0:
 	get:
@@ -15,11 +18,10 @@ class_name VirtualCamera extends Node3D
 	set(value):
 		if priority == value: return
 		priority = value
+		if _isEditorMode: return
 		VirtualCameraService.onPriorityModified.emit(self)
-		print(name, " ", "priority ",value)
 
 @export var trackingTarget : Node3D
-
 @export var lookAt : Node3D
 
 @export var processMethod: UtilsCamera.UpdateMethods = UtilsCamera.UpdateMethods.DEFAULT_PROCESS:
@@ -27,12 +29,19 @@ class_name VirtualCamera extends Node3D
 		return processMethod
 	set(value):
 		processMethod = value
-		print(name, " ", "processMethod ",value)
-	
+
+@export var lens : LensComponent
+
 func _enter_tree():
+	if _isEditorMode: return
 	VirtualCameraService.addVirtualCamera(self)
 
+func _ready():
+	if not _isEditorMode: return
+	UtilsCamera.createLensComponent(self)	
+	
 func _exit_tree():
+	if _isEditorMode: return
 	VirtualCameraService.removeVirtualCamera(self)
 
 func _to_string():
